+++ CHANGELOG.md
@@ -30,6 +30,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * New `divergent()` revset function for divergent changes.
 
+* A new config option `remotes.<name>.auto-track-created-bookmarks` behaves
+  similarly to `auto-track-bookmarks`, but it only applies to bookmarks created
+  locally. Setting it to `"*"` is now the closest replacement for the deprecated
+  `git.push-new-bookmarks` option.
+
 ### Fixed bugs
 
 * `jj git push` now ensures that tracked remote bookmarks are updated even if